USBtoSerial: Add monitor at 1200 baud
Add little monitor which allows instant switch to boot-loader and
drive PF7 (which is A0 on the Arduino Micro) pin (to reset a slave
board, for example...)
To enter monitor mode, change baud-rate to 1200 baud. Every character
that is written in this mode will be consumed by the monitor program
and will _NOT_ be sent to the serial port (keep that in mind).