USBtoSerial: Add target 'program' and program.sh
The target 'program' calls 'program.sh' which can be used to program
Arduino Micro board with this enhanced version of USBtoSerial. The
script automatically switches (only on Linux) to monitor mode (by
changing baud-rate to 1200) and then enters the bootloader. Afterwards
the board will be flashed by using the 'avrdude' target and the
baud-rate will be reset to a default rate.